Separando novamente o arquivo gl.h, mas mantendo a nova estrutura
