+++ include/gl/init.h
@@ -1,0 +1,6 @@
+#ifndef S3D_GL_INIT_H
+#define S3D_GL_INIT_H
+
+void s3d_gl_init(const int width, const int height, const char *wndCaption);
+
+#endif
