+++ include/gl/load_buf.h
@@ -1,0 +1,6 @@
+#ifndef S3D_GL_LOAD_BUF_H
+#define S3D_GL_LOAD_BUF_H
+
+unsigned s3d_gl_load_buf(int type, const char *path, int hint);
+
+#endif
