+++ include/gl/load_buf_fmt.h
@@ -1,0 +1,6 @@
+#ifndef S3D_GL_LOAD_BUF_FMT_H
+#define S3D_GL_LOAD_BUF_FMT_H
+
+void s3d_gl_load_buf_fmt(const char *path);
+
+#endif
