+++ include/gl/load_dds.h
@@ -1,0 +1,6 @@
+#ifndef S3D_GL_LOAD_DDS_H
+#define S3D_GL_LOAD_DDS_H
+
+void s3d_gl_load_dds(unsigned *tex_id, const char *path);
+
+#endif
