+++ include/gl/should_close.h
@@ -1,0 +1,6 @@
+#ifndef S3D_GL_SHOULD_CLOSE_H
+#define S3D_GL_SHOULD_CLOSE_H
+
+int s3d_gl_should_close();
+
+#endif
