+++ include/gl/swap.h
@@ -1,0 +1,6 @@
+#ifndef S3D_GL_SWAP_H
+#define S3D_GL_SWAP_H
+ 
+void s3d_gl_swap();
+
+#endif
