+++ include/gl/window.h
@@ -1,0 +1,6 @@
+#ifndef S3D_GL_WINDOW_H
+#define S3D_GL_WINDOW_H
+
+extern struct GLFWwindow *s3d_gl_window;
+
+#endif
