+++ src/gl/load_buf.c
@@ -1,0 +1,26 @@
+#include "gl/load_buf.h"
+
+#include<GL/GLee.h>
+
+#include "err.h"
+#include "sys.h"
+
+unsigned s3d_gl_load_buf(int type, const char *path, int hint) {
+    unsigned buf_id;
+
+    glGenBuffers(1, &buf_id);
+
+    int len;
+    char *data = s3d_read_file(path, 0, &len);
+
+    glBindBuffer(type, buf_id);
+    glBufferData(type, len, data, hint);
+
+    assert_gl_ok();
+
+    free(data);
+
+    return buf_id;
+}
+
+
