+++ src/gl/window.c
@@ -1,0 +1,6 @@
+#include "gl/window.h"
+
+#include<GLFW/glfw3.h>
+
+GLFWwindow *s3d_gl_window = 0;
+
